+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 * Added Formstone.Ready handler to Core
 * Added additional AJAX parameters in Upload events
 * Fixed issue with moving input elements in Checkbox
+* Fixed issue with load order and `setup` calls
 * Removed IE 8 & 9 support
 
 ### 1.3.3